Recommend adding specs folder exclude list in `.solargraph.yml`
During the testing of the gem with the
[gitlabs](https://gitlab.com/gitlab-org/gitlab/-/blob/master/.solargraph.yml.example?ref_type=heads)
codebase, it resulted that there was a significant slow-down of the
solargraph due to specs pre-parsing at language server boot time.

Turns out, Solargraph treats `exclude:` key in `.solargraph.yml` config
file as "exclude pre-indexing" of directory, which in our case is
exactly what we want, being that specs ought to be isolated from one
another and the only dependency should be actually source code & test
utils.