Fix upcoming spec incompatibility with solargraph ApiMap
In an upcoming solargraph PR (see
https://github.com/castwide/solargraph/pull/1107 as an example), we
won't be able to rely on resolving RSpec constants whose gem names we
don't provide using the 'require' convention.

That said, we also should not be caching as aggressively, either - so
I'd recommend dropping this spec workaround, which seems to resolve
the issue seen in the PR above.